fix test execution
